tests/submission/PhilipTaylor/canvas/security.dataURI.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: security.dataURI</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.security.html">security</a>.dataURI</h1>
<p class="desc">data: URIs do not count as different-origin, and do not taint the canvas</p>

<div class="refs">Spec references:
<ul>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("data: URIs do not count as different-origin, and do not taint the canvas");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
var data = canvas.toDataURL();
ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
var img = new Image();
deferTest();
img.onload = wrapFunction(function ()
{
    ctx.drawImage(img, 0, 0);
    canvas.toDataURL(); // should be permitted
    _assertPixel(canvas, 50,25, 0,255,0,255, "50,25", "0,255,0,255");
});
img.src = data;


});
</script>

<div id="log"></div>
</body></html>
